requirements.txt
fastapi~=0.136.0
uvicorn
pyyaml~=6.0.3
itsdangerous
loguru
apscheduler
httpx

starlette~=1.0.0

# dev
pytest